fix: missing injectHtml options
