+++ src/transform/transformer.ts
@@ -5,6 +5,9 @@ import { VueCliTransformer } from './transformVuecli';
 import { WebpackTransformer } from './transformWebpack';
 import { recordConver } from '../utils/report'
 import type { AstParsingResult } from '../ast-parse/astParse';
+import type { WebpackPluginInstance } from 'webpack';
+import type { InjectOptions } from '../config/config';
+import { serializeObject } from '../generate/render';
 
 /**
  * general implementation for vue.config.js and webpack.config.js
@@ -96,4 +99,62 @@ export function transformImporters (context: TransformContext, astParsingResult?
   plugins.push(new RawValue('injectHtml()'))
 
   context.config.plugins = plugins
+}
+
+export function transformWebpackHtmlPlugin (htmlPlugin: WebpackPluginInstance, injectHtmlPluginOption: InjectOptions, data) {
+  if (htmlPlugin && htmlPlugin.options) {
+    // injectData
+    Object.keys(htmlPlugin.options).forEach(key => {
+      if ((key === 'title' || key === 'favicon') && htmlPlugin.options[key]) {
+        data[key] = htmlPlugin.options[key]
+      }
+    })
+    if (htmlPlugin.options?.templateParameters) {
+      Object.assign(data, htmlPlugin.options.templateParameters)
+    }
+    if (htmlPlugin.options?.meta) {
+      injectHtmlPluginOption.tags = []
+      Object.keys(htmlPlugin.options.meta).forEach(key => {
+        if (htmlPlugin.options.meta[key]) {
+          injectHtmlPluginOption.tags.push({
+            tag: 'meta',
+            attrs: {
+              name: key,
+              content: htmlPlugin.options.meta[key],
+              injectTo: 'head'
+            }
+          })
+        }
+      })
+    }
+    this.context.config.plugins = this.context.config.plugins || []
+    const injectHtmlPluginIndex = this.context.config.plugins.findIndex(p => p.value === 'injectHtml()')
+    if (injectHtmlPluginIndex >= 0) {
+      this.context.config.plugins[injectHtmlPluginIndex] = new RawValue('injectHtml(' + serializeObject(injectHtmlPluginOption, '    ') + ')')
+    } else {
+      this.context.config.plugins.push(new RawValue('injectHtml(' + serializeObject(injectHtmlPluginOption, '    ') + ')'))
+    }
+    if (this.context.importers.findIndex(importer => importer.key === 'vite-plugin-html') < 0) {
+      this.context.importers.push({
+        key: 'vite-plugin-html',
+        value: 'import { injectHtml } from \'vite-plugin-html\';'
+      })
+    }
+    // minify
+    if (htmlPlugin.options?.minify) {
+      const vitePluginHtmlImporterIndex = this.context.importers.findIndex(importer => importer.key === 'vite-plugin-html')
+      if (vitePluginHtmlImporterIndex >= 0) {
+        const minifyHtmlImporter = 'import { injectHtml, minifyHtml } from \'vite-plugin-html\';'
+        this.context.importers[vitePluginHtmlImporterIndex].value = minifyHtmlImporter
+      } else {
+        this.context.importers.push({
+          key: 'vite-plugin-html',
+          value: 'import { minifyHtml } from \'vite-plugin-html\';'
+        })
+      }
+      this.context.config.plugins = this.context.config.plugins || []
+      this.context.config.plugins.push(new RawValue('minifyHtml(' + serializeObject(htmlPlugin.options.minify, '    ') + ')'))
+    }
+  }
+  injectHtmlPluginOption.data = data
 }
